update plugins, keep plugins/core only
