+++ megatron/core/datasets/blended_dataset.py
@@ -16,7 +16,7 @@ from megatron.core.datasets.megatron_dataset import MegatronDataset
 from megatron.core.datasets.utils import normalize
 from megatron.core.utils import log_single_rank
 
-from plugin.core.datasets.utils import is_built_on_zero_rank
+from plugin.utils import is_built_on_zero_rank
 
 logger = logging.getLogger(__name__)
