+++ plugin/core/distributed/finalize_model_grads.py
@@ -22,7 +22,7 @@ from megatron.core.utils import (
 from megatron.core.distributed.finalize_model_grads import _get_main_grad_attr, _unshard_if_dtensor, _reshard_if_dtensor
 
 
-from plugin.hetero.p2p_communication import get_device_type_for_comm
+from plugin.utils import get_device_type_for_comm
 from plugin.decorators import plugin_implementation
 
 logger = logging.getLogger(__name__)
